+++ src/app/vendors/VendorItems.tsx
@@ -5,6 +5,7 @@ import { VENDORS } from 'app/search/d2-known-values';
 import { chainComparator, compareBy } from 'app/utils/comparators';
 import { uniqBy } from 'app/utils/util';
 import deprecatedMods from 'data/d2/deprecated-mods.json';
+import focusingItemOutputs from 'data/d2/focusing-item-outputs.json';
 import rahoolMats from 'data/d2/spider-mats.json';
 import _ from 'lodash';
 import { Link } from 'react-router-dom';
@@ -194,7 +195,9 @@ export default function VendorItems({
                             key={vendorItem.key}
                             item={vendorItem}
                             owned={Boolean(
-                              ownedItemHashes?.has(vendorItem.item.hash) || vendorItem.owned
+                              ownedItemHashes?.has(vendorItem.item.hash) ||
+                                vendorItem.owned ||
+                                ownedItemHashes?.has(focusingItemOutputs[vendorItem.item.hash])
                             )}
                             characterId={characterId}
                           />