+++ docs/CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Next
 
+* Items in the "Focused Decoding" and "Legacy Gear" screens within the Vendors page will now correctly show collection and inventory checkmarks.
+
 ## 7.61.0 <span class="changelog-date">(2023-03-19)</span>
 
 * Hashtags for items and loadouts can now contain emoji.